Org reader: fix absolute file schema links
Org-mode allows links like "file:/absolute.link" which are exported to
"file:///absolute.link" URIs (and this is the default number of slashes
created by org-insert-link). In fact, it allows any number of leading
slashes in these links and normalizes to three slashes during export, so
we do the same.